production/base_function/requirements.txt
gevent==21.12.0
asyncio==3.4.3
aiohttp==3.8.1
opentelemetry-api==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-instrumentation-redis==0.30b1
opentelemetry.instrumentation.fastapi==0.30b1
opentelemetry-instrumentation-aiohttp-client==0.30b1
opentelemetry-exporter-otlp-proto-http==1.11.1
redis==3.5.3
rejson==0.5.4
Werkzeug==1.0.1
six==1.12.0
fastapi==0.78.0
uvicorn==0.18.1
python-multipart==0.0.5
python-dateutil==2.8.1
prometheus-fastapi-instrumentator==5.7.1
coverage==5.5
protobuf==3.19.1